Refactor GenerateTestModal to use ModalShell
